adding environment.hpp 2
