+++ include/cpplox/interpreter.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cpplox/environment.hpp>
 #include <cpplox/error.hpp>
 #include <cpplox/expression.hpp>
 #include <cpplox/statement.hpp>
@@ -18,6 +19,8 @@ class Interpreter
 public:
   auto execute(const std::vector<Stmt> & program) -> std::optional<RuntimeError>;
 };
+
+auto evaluate_expr(const Expr & expr) -> std::variant<Value, RuntimeError>;
 
 }  // namespace interpreter
 }  // namespace lox
