+++ test/test_expression.cpp
@@ -1,3 +1,4 @@
+#include <cpplox/interpreter.hpp>
 #include <cpplox/parser.hpp>
 #include <cpplox/tokenizer.hpp>
 #include <cpplox/variant.hpp>
